# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: 247URL API
  version: 1.1.0
  description: Versioned API for authentication, link management, analytics, QR codes, moderation, and account management.
  license: { name: MIT, identifier: MIT }
servers:
  - url: /api/v1
    description: Same-host API base
...
security: [{ sanctum: [] }]
paths:
...
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Authentication]
      security: []
      summary: Reset a password and revoke existing tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, email, password, password_confirmation]
              properties:
                token: { type: string }
                email: { type: string, format: email }
                password: { type: string, format: password }
                password_confirmation: { type: string, format: password }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/email/verification-notification:
    post:
      operationId: resendVerification
      tags: [Authentication]
      summary: Resend email verification for an active account
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    sanctum: { type: http, scheme: bearer, bearerFormat: Sanctum }
...
  responses:
    Success: { description: Successful operation }
...
    Unauthorized: { description: Authentication required }
    Forbidden: { description: Authenticated but not permitted }
...
    RateLimited: { description: Rate limit exceeded }
    ValidationError:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } }
  schemas:
...
    ValidationError:
      type: object
      properties: { message: { type: string, example: Validation failed }, errors: { type: object, additionalProperties: { type: array, items: { type: string } } } }